Resolve the locals call in consants with more readable code
The previosu solution was kind of leaning into code golf, it was hard to read what was going on and the comments weren't too helpful either. Especially since this isn't even part of the core functionality, it shouldn't have been written like this. A much more simple solution that's much clearer and only like one line longer has been implemented.